Add `apply_flags` method for test flags (#11954)
- Add TestFlags class base, add `apply_flags` method to handle applying test flags to arguments.
- Refactored and update redundant code after adding the `apply_flags` method.
- Added missing `on_device` parameter to `test_frontend_method`.
- Updated `create_args_kwargs` to make use of the `on_device` parameter.
