+++ ivy_tests/test_ivy/test_frontends/test_torch/test_tensor.py
@@ -139,6 +139,7 @@ def test_torch_instance_chunk(
     frontend_method_data,
     init_flags,
     method_flags,
+    on_device,
 ):
     input_dtype, x, dim = dtype_x_dim
     helpers.test_frontend_method(
@@ -155,6 +156,7 @@ def test_torch_instance_chunk(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -180,6 +182,7 @@ def test_torch_instance_any(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x, axis = dtype_input_axis
     helpers.test_frontend_method(
@@ -196,6 +199,7 @@ def test_torch_instance_any(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -221,6 +225,7 @@ def test_torch_instance_all(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x, axis = dtype_input_axis
     helpers.test_frontend_method(
@@ -237,6 +242,7 @@ def test_torch_instance_all(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -261,6 +267,7 @@ def test_torch_instance_add(
     frontend_method_data,
     init_flags,
     method_flags,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -278,6 +285,7 @@ def test_torch_instance_add(
         method_flags=method_flags,
         frontend=frontend,
         atol_=1e-02,
+        on_device=on_device,
     )
 
 
@@ -325,6 +333,7 @@ def test_torch_instance_new_ones(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -372,6 +381,7 @@ def test_torch_instance_new_zeros(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -396,6 +406,7 @@ def test_torch_instance_reshape(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_x
     shape = {
@@ -418,6 +429,7 @@ def test_torch_instance_reshape(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -436,6 +448,7 @@ def test_torch_instance_reshape_as(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_x
     helpers.test_frontend_method(
@@ -451,6 +464,7 @@ def test_torch_instance_reshape_as(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -470,6 +484,7 @@ def test_torch_instance_sin(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -483,6 +498,7 @@ def test_torch_instance_sin(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -502,6 +518,7 @@ def test_torch_instance_arcsin(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -515,6 +532,7 @@ def test_torch_instance_arcsin(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -535,6 +553,7 @@ def test_torch_instance_sum(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -548,6 +567,7 @@ def test_torch_instance_sum(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -567,6 +587,7 @@ def test_torch_instance_atan(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -580,6 +601,7 @@ def test_torch_instance_atan(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -599,6 +621,7 @@ def test_torch_instance_atan2(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -614,6 +637,7 @@ def test_torch_instance_atan2(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -633,6 +657,7 @@ def test_torch_instance_sin_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -646,6 +671,7 @@ def test_torch_instance_sin_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -665,6 +691,7 @@ def test_torch_instance_cos(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -678,6 +705,7 @@ def test_torch_instance_cos(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -697,6 +725,7 @@ def test_torch_instance_cos_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -710,6 +739,7 @@ def test_torch_instance_cos_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -729,6 +759,7 @@ def test_torch_instance_sinh(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -742,6 +773,7 @@ def test_torch_instance_sinh(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -761,6 +793,7 @@ def test_torch_instance_sinh_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -774,6 +807,7 @@ def test_torch_instance_sinh_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -793,6 +827,7 @@ def test_torch_instance_cosh(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -806,6 +841,7 @@ def test_torch_instance_cosh(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -825,6 +861,7 @@ def test_torch_instance_cosh_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -838,6 +875,7 @@ def test_torch_instance_cosh_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -861,6 +899,7 @@ def test_torch_instance_view(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_x
     helpers.test_frontend_method(
@@ -876,6 +915,7 @@ def test_torch_instance_view(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -893,6 +933,7 @@ def test_torch_instance_float(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_x
     helpers.test_frontend_method(
@@ -906,6 +947,7 @@ def test_torch_instance_float(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -925,6 +967,7 @@ def test_torch_instance_asinh(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -940,6 +983,7 @@ def test_torch_instance_asinh(
         frontend=frontend,
         rtol_=1e-2,
         atol_=1e-2,
+        on_device=on_device,
     )
 
 
@@ -959,6 +1003,7 @@ def test_torch_instance_asinh_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -974,6 +1019,7 @@ def test_torch_instance_asinh_(
         frontend=frontend,
         rtol_=1e-2,
         atol_=1e-2,
+        on_device=on_device,
     )
 
 
@@ -993,6 +1039,7 @@ def test_torch_instance_tan(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1006,6 +1053,7 @@ def test_torch_instance_tan(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1025,6 +1073,7 @@ def test_torch_instance_tanh(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1038,6 +1087,7 @@ def test_torch_instance_tanh(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1057,6 +1107,7 @@ def test_torch_instance_tanh_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1070,6 +1121,7 @@ def test_torch_instance_tanh_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1089,6 +1141,7 @@ def test_torch_instance_asin(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1102,6 +1155,7 @@ def test_torch_instance_asin(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1120,6 +1174,7 @@ def test_torch_instance_amax(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_x
     helpers.test_frontend_method(
@@ -1133,6 +1188,7 @@ def test_torch_instance_amax(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1151,6 +1207,7 @@ def test_torch_instance_abs(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1164,6 +1221,7 @@ def test_torch_instance_abs(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1182,6 +1240,7 @@ def test_torch_instance_abs_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1195,6 +1254,7 @@ def test_torch_instance_abs_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1213,6 +1273,7 @@ def test_torch_instance_amin(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_x
     helpers.test_frontend_method(
@@ -1226,6 +1287,7 @@ def test_torch_instance_amin(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1246,6 +1308,7 @@ def test_torch_instance_aminmax(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_x
     helpers.test_frontend_method(
@@ -1261,6 +1324,7 @@ def test_torch_instance_aminmax(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1280,6 +1344,7 @@ def test_torch_instance_contiguous(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1293,6 +1358,7 @@ def test_torch_instance_contiguous(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1312,6 +1378,7 @@ def test_torch_instance_log(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1325,6 +1392,7 @@ def test_torch_instance_log(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1344,6 +1412,7 @@ def test_torch_instance_log_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1357,6 +1426,7 @@ def test_torch_instance_log_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1379,6 +1449,7 @@ def test_torch_special_add(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1394,6 +1465,7 @@ def test_torch_special_add(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1415,6 +1487,7 @@ def test_torch_special_long(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1428,6 +1501,7 @@ def test_torch_special_long(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1450,6 +1524,7 @@ def test_torch_special_radd(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1465,6 +1540,7 @@ def test_torch_special_radd(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1487,6 +1563,7 @@ def test_torch_special_sub(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1502,6 +1579,7 @@ def test_torch_special_sub(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1524,6 +1602,7 @@ def test_torch_special_mul(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1539,6 +1618,7 @@ def test_torch_special_mul(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1558,6 +1638,7 @@ def test_torch_special_rsub(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1573,6 +1654,7 @@ def test_torch_special_rsub(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1595,6 +1677,7 @@ def test_torch_special_rmul(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1610,6 +1693,7 @@ def test_torch_special_rmul(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1633,6 +1717,7 @@ def test_torch_special_truediv(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1648,6 +1733,7 @@ def test_torch_special_truediv(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1690,6 +1776,7 @@ def test_torch_instance_to(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x, method_num_positional_args, method_all_as_kwargs_np = args_kwargs
     method_flags.num_positional_args = method_num_positional_args
@@ -1704,6 +1791,7 @@ def test_torch_instance_to(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1723,6 +1811,7 @@ def test_torch_instance_arctan(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1736,6 +1825,7 @@ def test_torch_instance_arctan(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1755,6 +1845,7 @@ def test_torch_instance_arctan_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1768,6 +1859,7 @@ def test_torch_instance_arctan_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1787,6 +1879,7 @@ def test_torch_instance_arctan2(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1802,6 +1895,7 @@ def test_torch_instance_arctan2(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1821,6 +1915,7 @@ def test_torch_instance_arctan2_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1836,6 +1931,7 @@ def test_torch_instance_arctan2_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1855,6 +1951,7 @@ def test_torch_instance_acos(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1868,6 +1965,7 @@ def test_torch_instance_acos(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1887,6 +1985,7 @@ def test_torch_instance_new_tensor(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -1903,6 +2002,7 @@ def test_torch_instance_new_tensor(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -1974,6 +2074,7 @@ def test_torch_instance_getitem(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     data = x[0]
@@ -1987,6 +2088,7 @@ def test_torch_instance_getitem(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2007,6 +2109,7 @@ def test_torch_instance_view_as(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_x
     helpers.test_frontend_method(
@@ -2022,6 +2125,7 @@ def test_torch_instance_view_as(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2047,6 +2151,7 @@ def test_torch_instance_unsqueeze(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_value
     helpers.test_frontend_method(
@@ -2062,6 +2167,7 @@ def test_torch_instance_unsqueeze(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2087,6 +2193,7 @@ def test_torch_instance_unsqueeze_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_value
     helpers.test_frontend_method(
@@ -2102,6 +2209,7 @@ def test_torch_instance_unsqueeze_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2131,6 +2239,7 @@ def test_torch_instance_split(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_value
     helpers.test_frontend_method(
@@ -2147,6 +2256,7 @@ def test_torch_instance_split(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2168,6 +2278,7 @@ def test_torch_instance_vsplit(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_value
     helpers.test_frontend_method(
@@ -2181,6 +2292,7 @@ def test_torch_instance_vsplit(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2202,6 +2314,7 @@ def test_torch_instance_hsplit(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_value
     # TODO: remove the assumption when these bugfixes are merged and version-pinned
@@ -2223,6 +2336,7 @@ def test_torch_instance_hsplit(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2244,6 +2358,7 @@ def test_torch_instance_dsplit(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_value
     helpers.test_frontend_method(
@@ -2257,6 +2372,7 @@ def test_torch_instance_dsplit(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2275,6 +2391,7 @@ def test_torch_instance_detach(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -2288,6 +2405,7 @@ def test_torch_instance_detach(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2306,6 +2424,7 @@ def test_torch_instance_dim(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -2319,6 +2438,7 @@ def test_torch_instance_dim(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2337,6 +2457,7 @@ def test_torch_instance_ndimension(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -2350,6 +2471,7 @@ def test_torch_instance_ndimension(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2410,6 +2532,7 @@ def test_torch_instance_new_full(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -2426,6 +2549,7 @@ def test_torch_instance_new_full(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2449,6 +2573,7 @@ def test_torch_instance_new_empty(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -2464,6 +2589,7 @@ def test_torch_instance_new_empty(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2503,6 +2629,7 @@ def test_torch_instance_expand(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x, shape = dtype_x_shape
     if unpack_shape:
@@ -2527,6 +2654,7 @@ def test_torch_instance_expand(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2545,6 +2673,7 @@ def test_torch_instance_expand_as(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_x
     helpers.test_frontend_method(
@@ -2560,6 +2689,7 @@ def test_torch_instance_expand_as(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2607,6 +2737,7 @@ def test_torch_instance_unfold(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x, axis, size, step = dtype_values_args
     print(axis, size, step)
@@ -2625,6 +2756,7 @@ def test_torch_instance_unfold(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2644,6 +2776,7 @@ def test_torch_special_mod(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -2659,6 +2792,7 @@ def test_torch_special_mod(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2677,6 +2811,7 @@ def test_torch_instance_long(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -2690,6 +2825,7 @@ def test_torch_instance_long(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2708,6 +2844,7 @@ def test_torch_instance_max(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_x
     helpers.test_frontend_method(
@@ -2721,6 +2858,7 @@ def test_torch_instance_max(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2753,6 +2891,7 @@ def test_torch_instance_is_cuda(
     frontend_method_data,
     init_flags,
     method_flags,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     device = "cpu" if device is False else "gpu:0"
@@ -2771,6 +2910,7 @@ def test_torch_instance_is_cuda(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2790,6 +2930,7 @@ def test_torch_instance_bitwise_not(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -2803,6 +2944,7 @@ def test_torch_instance_bitwise_not(
         method_flags=method_flags,
         method_all_as_kwargs_np={},
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2822,6 +2964,7 @@ def test_torch_instance_bitwise_and(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -2837,6 +2980,7 @@ def test_torch_instance_bitwise_and(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2856,6 +3000,7 @@ def test_torch_instance_bitwise_or(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -2871,6 +3016,7 @@ def test_torch_instance_bitwise_or(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2890,6 +3036,7 @@ def test_torch_instance_add_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -2905,6 +3052,7 @@ def test_torch_instance_add_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2925,6 +3073,7 @@ def test_torch_instance_arccos_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -2938,6 +3087,7 @@ def test_torch_instance_arccos_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2958,6 +3108,7 @@ def test_torch_instance_arccos(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -2971,6 +3122,7 @@ def test_torch_instance_arccos(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -2991,6 +3143,7 @@ def test_torch_instance_acos_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -3004,6 +3157,7 @@ def test_torch_instance_acos_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3024,6 +3178,7 @@ def test_torch_instance_asin_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -3037,6 +3192,7 @@ def test_torch_instance_asin_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3057,6 +3213,7 @@ def test_torch_instance_arcsin_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -3070,6 +3227,7 @@ def test_torch_instance_arcsin_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3089,6 +3247,7 @@ def test_torch_instance_atan_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -3102,6 +3261,7 @@ def test_torch_instance_atan_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3121,6 +3281,7 @@ def test_torch_instance_tan_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -3134,6 +3295,7 @@ def test_torch_instance_tan_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3154,6 +3316,7 @@ def test_torch_instance_atanh(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -3167,6 +3330,7 @@ def test_torch_instance_atanh(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3187,6 +3351,7 @@ def test_torch_instance_atanh_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -3200,6 +3365,7 @@ def test_torch_instance_atanh_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3220,6 +3386,7 @@ def test_torch_instance_arctanh(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -3253,6 +3420,7 @@ def test_torch_instance_arctanh_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -3266,6 +3434,7 @@ def test_torch_instance_arctanh_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3288,6 +3457,7 @@ def test_torch_instance_pow(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     dtype = input_dtype[0]
@@ -3306,6 +3476,7 @@ def test_torch_instance_pow(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3325,6 +3496,7 @@ def test_torch_instance_pow_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     dtype = input_dtype[0]
@@ -3343,6 +3515,7 @@ def test_torch_instance_pow_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3409,6 +3582,7 @@ def test_torch_instance_argmax(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x, axis = dtype_input_axis
     helpers.test_frontend_method(
@@ -3425,6 +3599,7 @@ def test_torch_instance_argmax(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3454,6 +3629,7 @@ def test_torch_instance_argmin(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x, axis = dtype_input_axis
     helpers.test_frontend_method(
@@ -3470,6 +3646,7 @@ def test_torch_instance_argmin(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3496,6 +3673,7 @@ def test_torch_instance_argsort(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x, axis = dtype_input_axis
     helpers.test_frontend_method(
@@ -3512,6 +3690,7 @@ def test_torch_instance_argsort(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3530,6 +3709,7 @@ def test_torch_instance_ceil(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -3543,6 +3723,7 @@ def test_torch_instance_ceil(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3561,6 +3742,7 @@ def test_torch_instance_argwhere(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -3574,6 +3756,7 @@ def test_torch_instance_argwhere(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3598,6 +3781,7 @@ def test_torch_instance_size(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -3613,6 +3797,7 @@ def test_torch_instance_size(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3631,6 +3816,7 @@ def test_torch_instance_min(
     frontend_method_data,
     init_flags,
     method_flags,
+    on_device,
 ):
     input_dtype, x = dtype_x
     helpers.test_frontend_method(
@@ -3644,6 +3830,7 @@ def test_torch_instance_min(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3670,6 +3857,7 @@ def test_torch_instance_matmul(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     dtype, tensor1, tensor2 = dtype_tensor1_tensor2
     helpers.test_frontend_method(
@@ -3683,6 +3871,7 @@ def test_torch_instance_matmul(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3721,6 +3910,7 @@ def test_torch_instance_permute(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     x, idxes, dtype = dtype_values_axis
     unpack_dims = True
@@ -3746,6 +3936,7 @@ def test_torch_instance_permute(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3766,6 +3957,7 @@ def test_torch_instance_mean(
     frontend_method_data,
     init_flags,
     method_flags,
+    on_device,
 ):
     input_dtype, x = dtype_x
     helpers.test_frontend_method(
@@ -3779,6 +3971,7 @@ def test_torch_instance_mean(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3810,6 +4003,7 @@ def test_torch_instance_transpose(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_value
     helpers.test_frontend_method(
@@ -3823,6 +4017,7 @@ def test_torch_instance_transpose(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3854,6 +4049,7 @@ def test_torch_instance_transpose_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_value
     helpers.test_frontend_method(
@@ -3870,6 +4066,7 @@ def test_torch_instance_transpose_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3901,6 +4098,7 @@ def test_torch_instance_flatten(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     if start_dim > end_dim:
         temp = start_dim
@@ -3921,6 +4119,7 @@ def test_torch_instance_flatten(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3948,6 +4147,7 @@ def test_torch_instance_cumsum(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_value
     helpers.test_frontend_method(
@@ -3964,6 +4164,7 @@ def test_torch_instance_cumsum(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -3989,6 +4190,7 @@ def test_torch_instance_cumsum_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_value
     helpers.test_frontend_method(
@@ -4005,6 +4207,7 @@ def test_torch_instance_cumsum_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4032,6 +4235,7 @@ def test_torch_instance_sort(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_value
     helpers.test_frontend_method(
@@ -4048,6 +4252,7 @@ def test_torch_instance_sort(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4066,6 +4271,7 @@ def test_torch_instance_sigmoid(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_x
     helpers.test_frontend_method(
@@ -4079,6 +4285,7 @@ def test_torch_instance_sigmoid(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4103,6 +4310,7 @@ def test_torch_instance_softmax(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x, axis = dtype_x_and_axis
     helpers.test_frontend_method(
@@ -4119,6 +4327,7 @@ def test_torch_instance_softmax(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4156,6 +4365,7 @@ def test_torch_instance_repeat(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x, repeats = dtype_x_repeats
     repeat = {
@@ -4176,6 +4386,7 @@ def test_torch_instance_repeat(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4197,6 +4408,7 @@ def test_torch_instance_unbind(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtypes, x, axis = dtype_value_axis
     helpers.test_frontend_method(
@@ -4212,6 +4424,7 @@ def test_torch_instance_unbind(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4234,6 +4447,7 @@ def test_torch_special_eq(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -4249,6 +4463,7 @@ def test_torch_special_eq(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4268,6 +4483,7 @@ def test_torch_instance_inverse(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -4281,6 +4497,7 @@ def test_torch_instance_inverse(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4302,6 +4519,7 @@ def test_torch_instance_neg(
     frontend_method_data,
     init_flags,
     method_flags,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -4315,6 +4533,7 @@ def test_torch_instance_neg(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4333,6 +4552,7 @@ def test_torch_instance_int(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -4346,6 +4566,7 @@ def test_torch_instance_int(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4364,6 +4585,7 @@ def test_torch_instance_bool(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -4377,6 +4599,7 @@ def test_torch_instance_bool(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4397,6 +4620,7 @@ def test_torch_instance_type(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -4412,6 +4636,7 @@ def test_torch_instance_type(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4431,6 +4656,7 @@ def test_torch_instance_type_as(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -4446,6 +4672,7 @@ def test_torch_instance_type_as(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4464,6 +4691,7 @@ def test_torch_instance_byte(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -4477,6 +4705,7 @@ def test_torch_instance_byte(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4499,6 +4728,7 @@ def test_torch_instance_ne(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -4514,6 +4744,7 @@ def test_torch_instance_ne(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4535,6 +4766,7 @@ def test_torch_instance_squeeze(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_value
     helpers.test_frontend_method(
@@ -4550,6 +4782,7 @@ def test_torch_instance_squeeze(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4568,6 +4801,7 @@ def test_torch_instance_flip(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     x, idxes, dtype = dtype_values_axis
     helpers.test_frontend_method(
@@ -4583,6 +4817,7 @@ def test_torch_instance_flip(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4604,6 +4839,7 @@ def test_torch_instance_tril(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_values
     helpers.test_frontend_method(
@@ -4619,6 +4855,7 @@ def test_torch_instance_tril(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4637,6 +4874,7 @@ def test_torch_instance_sqrt(
     frontend_method_data,
     init_flags,
     method_flags,
+    on_device,
 ):
     input_dtype, x = dtype_x
     helpers.test_frontend_method(
@@ -4648,6 +4886,7 @@ def test_torch_instance_sqrt(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4669,6 +4908,7 @@ def test_torch_instance_index_select(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtypes, input, indices, axis, batch_dims = params_indices_others
     helpers.test_frontend_method(
@@ -4685,6 +4925,7 @@ def test_torch_instance_index_select(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4741,6 +4982,7 @@ def test_torch_instance_clamp(
     frontend_method_data,
     init_flags,
     method_flags,
+    on_device,
 ):
     input_dtype, x, min, max = dtype_and_x_min_max
     helpers.test_frontend_method(
@@ -4754,6 +4996,7 @@ def test_torch_instance_clamp(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4770,6 +5013,7 @@ def test_torch_instance_clamp_(
     frontend_method_data,
     init_flags,
     method_flags,
+    on_device,
 ):
     input_dtype, x, min, max = dtype_and_x_min_max
     helpers.test_frontend_method(
@@ -4783,6 +5027,7 @@ def test_torch_instance_clamp_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4805,6 +5050,7 @@ def test_torch_special_gt(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -4820,6 +5066,7 @@ def test_torch_special_gt(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4842,6 +5089,7 @@ def test_torch_special_ne(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -4857,6 +5105,7 @@ def test_torch_special_ne(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4879,6 +5128,7 @@ def test_torch_special_lt(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -4894,6 +5144,7 @@ def test_torch_special_lt(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4916,6 +5167,7 @@ def test_torch_special_or(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -4931,6 +5183,7 @@ def test_torch_special_or(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4947,6 +5200,7 @@ def test_torch_instance_where(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     cond, xs, dtypes = broadcastables
     helpers.test_frontend_method(
@@ -4963,6 +5217,7 @@ def test_torch_instance_where(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -4982,6 +5237,7 @@ def test_torch_instance_clone(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -4995,6 +5251,7 @@ def test_torch_instance_clone(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -5014,6 +5271,7 @@ def test_torch_special_invert(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -5027,6 +5285,7 @@ def test_torch_special_invert(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -5046,6 +5305,7 @@ def test_torch_instance_acosh(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -5059,6 +5319,7 @@ def test_torch_instance_acosh(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -5078,6 +5339,7 @@ def test_torch_instance_real(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -5091,6 +5353,7 @@ def test_torch_instance_real(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -5119,6 +5382,7 @@ def test_torch_instance_masked_fill(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     dtype, x, mask, val = x_mask_val
     helpers.test_frontend_method(
@@ -5135,6 +5399,7 @@ def test_torch_instance_masked_fill(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -5154,6 +5419,7 @@ def test_torch_instance_acosh_(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -5167,6 +5433,7 @@ def test_torch_instance_acosh_(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
 
 
@@ -5185,6 +5452,7 @@ def test_torch_instance_numpy(
     init_flags,
     method_flags,
     frontend,
+    on_device,
 ):
     input_dtype, x = dtype_and_x
     helpers.test_frontend_method(
@@ -5198,4 +5466,5 @@ def test_torch_instance_numpy(
         init_flags=init_flags,
         method_flags=method_flags,
         frontend=frontend,
+        on_device=on_device,
     )
